+++ test/VaultSharp.Samples/Program.cs
@@ -152,7 +152,7 @@ namespace VaultSharp.Samples
 
         private static void RunCubbyholeSamples()
         {
-            var secretPath = "test-path";
+            var secretPath = "a/test-path";
 
             var value = new Dictionary<string, object> { { "key1", "val1" }, { "key2", 2 } };
             _authenticatedVaultClient.V1.Secrets.Cubbyhole.WriteSecretAsync(secretPath, value).Wait();
@@ -166,8 +166,16 @@ namespace VaultSharp.Samples
             var read2 = _authenticatedVaultClient.V1.Secrets.Cubbyhole.ReadSecretAsync(secretPath).Result;
             Assert.True(read2.Data.Count == 3);
 
+            var value2 = new Dictionary<string, object> { { "key1", "val1" }, { "key2", 2 } };
+            _authenticatedVaultClient.V1.Secrets.Cubbyhole.WriteSecretAsync(secretPath + "2", value2).Wait();
+
+            var paths = _authenticatedVaultClient.V1.Secrets.Cubbyhole.ReadSecretPathsAsync("a/").Result;
+            Assert.True(paths.Data.Keys.Count() == 2);
+
             _authenticatedVaultClient.V1.Secrets.Cubbyhole.DeleteSecretAsync(secretPath).Wait();
             Assert.ThrowsAsync<VaultApiException>(() => _authenticatedVaultClient.V1.Secrets.Cubbyhole.ReadSecretAsync(secretPath)).Wait();
+
+            _authenticatedVaultClient.V1.Secrets.Cubbyhole.DeleteSecretAsync(secretPath + "2").Wait();
         }
 
         private static void RunTransitSamples()