+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   * Active Directory Secrets Engine: Add support for offering credentials.
   * AWS Secrets Engine: Add support for generating dynamic IAM credentials & STS IAM credentials.
   * Cubbhole Secrets Engine: Add support for read secret, read paths, write secret and delete secret APIs.
+  * Database Secrets Engine: Add support for generating dynamic DB credentials.
 
 ## 0.10.4001 (August 10, 2018)